# web/openapi/mengnex.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/manga/chapters/{id}/reader": {
      "get": {
        "tags": [
          "manga"
        ],
        "operationId": "get_reader",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MangaReaderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/manga/{id}": {
      "get": {
        "tags": [
          "manga"
        ],
        "operationId": "get_series",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MangaDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/media/files/{id}/content": {
      "get": {
        "tags": [
          "media"
        ],
        "operationId": "get_media_file_content",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Media file id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media file content"
          },
          "404": {
            "description": "Media file not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MangaChapterResponse": {
        "type": "object",
        "required": [
          "id",
          "title",
          "page_count"
        ],
        "properties": {
          "cover_src": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "page_count": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "MangaDetailResponse": {
        "type": "object",
        "required": [
          "id",
          "library_id",
          "title",
          "layout",
          "chapter_count",
          "page_count",
          "chapters"
        ],
        "properties": {
          "chapter_count": {
            "type": "integer",
            "format": "int64"
          },
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaChapterResponse"
            }
          },
          "cover_src": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "layout": {
            "type": "string"
          },
          "library_id": {
            "type": "string"
          },
          "page_count": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "MangaPageResponse": {
        "type": "object",
        "required": [
          "id",
          "src",
          "file_name",
          "page_number"
        ],
        "properties": {
          "file_name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "page_number": {
            "type": "integer",
            "format": "int64"
          },
          "src": {
            "type": "string"
          }
        }
      },
      "MangaReaderResponse": {
        "type": "object",
        "required": [
          "chapter_id",
          "series_id",
          "title",
          "pages"
        ],
        "properties": {
          "chapter_id": {
            "type": "string"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaPageResponse"
            }
          },
          "series_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}